Revert to a single solution, no settings
